Rearrange package
