+++ ipassign/__init__.py
@@ -1,10 +1,9 @@
-from .utils import acknowledgements, commands
-from .message import MAX_PACKET_LENGTH, Message, MIN_PACKET_LENGTH
-from .payload import Acknowledgement, Payload
+from .protocol.enums import acknowledgements, commands
+from .protocol.message import MAX_PACKET_LENGTH, Message, MIN_PACKET_LENGTH
+from .protocol.payload import Acknowledgement, Payload
 
-# UDP Multicast constants
-MULTICAST_ADDR = '225.0.0.37'
-MULTICAST_PORT = 12345
+from .networking import is_known_hostname, MULTICAST_ADDR, MULTICAST_PORT
 
-__all__ = [acknowledgements, Acknowledgement, commands, MAX_PACKET_LENGTH,
-           Message, MIN_PACKET_LENGTH, MULTICAST_ADDR, MULTICAST_PORT, Payload]
+__all__ = [acknowledgements, Acknowledgement, commands, is_known_hostname,
+           MAX_PACKET_LENGTH, Message, MIN_PACKET_LENGTH, MULTICAST_ADDR,
+           MULTICAST_PORT, Payload]
