+++ ipassign/protocol/enums.py
@@ -1,0 +1,36 @@
+from enum import Enum
+
+
+class commands(Enum):
+    REQUEST_CONFIG = 0x0002
+    SEND_CONFIG = 0x0003
+
+    RESET = 0x0004
+    RESET_ACK = 0x0005
+
+    CHANGE_IP = 0x0006
+    CHANGE_IP_ACK = 0x0007
+
+    SEND_RAW_DATA = 0x0008
+
+    SEND_FIRMWARE_DATA = 0x000A
+    WRITE_FIRMWARE = 0x000C
+
+    SEND_ASC_DATA = 0x000E
+
+    UPDATE_CONFIG = 0x000F
+    UPDATE_CONFIG_ACK = 0x0010
+
+
+class acknowledgements(Enum):
+    OK = 0x0000
+
+    ERR_SIZE_PACK = 0x0140
+    ERR_OPEN_SOCKET = 0x0141
+    ERR_CLEAR_FLAG = 0x0142
+    ERR_BAD_GW = 0x0143
+    ERR_SET_IP_BC_NM = 0x0145
+    ERR_SET_GW = 0x0146
+    ERR_SET_FLAG = 0x0147
+    ERR_SET_MCAST = 0x0148
+    ERR_SET_HOSTNAME = 0x0149
